add encryption to Activity Log, refactor modules
